Fixing deprecated libraries + code bugs

+++ src/main/java/org/linkeddatafragments/datasource/hdt/HdtBasedRequestProcessorForTPFs.java
@@ -112,9 +112,9 @@ public class HdtBasedRequestProcessorForTPFs
             // look up the result from the HDT datasource)
             
            // System.out.println("look up the result from the HDT datasource)");
-            int subjectId = subject.isVariable() ? 0 : dictionary.getIntID(subject.asConstantTerm().asNode(), TripleComponentRole.SUBJECT);
-            int predicateId = predicate.isVariable() ? 0 : dictionary.getIntID(predicate.asConstantTerm().asNode(), TripleComponentRole.PREDICATE);
-            int objectId = object.isVariable() ? 0 : dictionary.getIntID(object.asConstantTerm().asNode(), TripleComponentRole.OBJECT);
+            int subjectId = subject.isVariable() ? 0 : (int) dictionary.getIntID(subject.asConstantTerm().asNode(), TripleComponentRole.SUBJECT);
+            int predicateId = predicate.isVariable() ? 0 : (int) dictionary.getIntID(predicate.asConstantTerm().asNode(), TripleComponentRole.PREDICATE);
+            int objectId = object.isVariable() ? 0 : (int) dictionary.getIntID(object.asConstantTerm().asNode(), TripleComponentRole.OBJECT);
         
             if (subjectId < 0 || predicateId < 0 || objectId < 0) {
                 return createEmptyTriplePatternFragment();
